Abstract out db ref table class methods.
